# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: PetMatch API
  description: API para sistema de adoção de animais
  version: 1.0.0
  contact:
    name: PetMatch Support
servers:
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento
...
paths:
  /v1/api/auth/register:
    post:
      tags:
        - Autenticação
      summary: Registrar novo usuário
      description: Cria um novo usuário no sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "usuario@exemplo.com"
              password: "Senha@123"
              name: "João Silva"
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Dados inválidos
        '409':
          description: Email já cadastrado
  /v1/api/auth/login:
    post:
      tags:
        - Autenticação
      summary: Login de usuário
      description: Autentica um usuário e retorna um token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "usuario@exemplo.com"
              password: "Senha@123"
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Credenciais inválidas
        '400':
          description: Dados inválidos
  /v1/api/auth/loginprova:
    post:
      tags:
        - Autenticação
      summary: Login de teste (DEV ONLY)
      description: Endpoint de teste que força credenciais fixas - REMOVER EM PRODUÇÃO
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
  /v1/api/eventos:
    get:
      tags:
        - Eventos
      summary: Listar todos os eventos
      description: Retorna lista com todos os eventos de adoção cadastrados
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Lista de eventos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventoResponseDto'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'

    post:
      tags:
        - Eventos
      summary: Criar novo evento
      description: |
        **Apenas ONGs podem criar eventos**
        
        Cria um novo evento de adoção. O ID da ONG é obtido automaticamente do token JWT.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarEventoDto'
            example:
              titulo: "Adoção em Belo Horizonte"
              descricao: "Grande evento de adoção com mais de 50 animais"
              dataEvento: "2025-12-10T10:00:00Z"
              local: "Praça da Estação, Belo Horizonte"
      responses:
        '201':
          description: Evento criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoResponseDto'
        '400':
          description: Requisição inválida (dados obrigatórios ausentes ou formato incorreto)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403':
          description: Usuário não possui permissão para criar eventos (apenas ONG)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'

    /v1/api/eventos/{id}:
      get:
        tags:
          - Eventos
        summary: Buscar evento por ID
        description: Retorna os detalhes de um evento específico baseado em seu ID
        security:
          - bearerAuth: [ ]
        parameters:
          - name: id
            in: path
            required: true
            description: ID único do evento (UUID)
            schema:
              type: string
              format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        responses:
          '200':
            description: Evento encontrado e retornado com sucesso
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/EventoResponseDto'
          '401':
            description: Token ausente ou inválido
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErroResponse'
          '404':
            description: Evento não encontrado
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErroResponse'
          '500':
            description: Erro interno do servidor
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErroResponse'

      delete:
        tags:
          - Eventos
        summary: Deletar evento por ID
        description: |
          **Apenas a ONG criadora pode deletar seu evento**

          Deleta um evento específico do sistema. O evento é excluído permanentemente.
        security:
          - bearerAuth: [ ]
        parameters:
          - name: id
            in: path
            required: true
            description: ID único do evento a ser deletado (UUID)
            schema:
              type: string
              format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        responses:
          '204':
            description: Evento deletado com sucesso (sem conteúdo na resposta)
          '401':
            description: Token ausente ou inválido
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErroResponse'
          '403':
            description: Usuário não possui permissão para deletar este evento
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErroResponse'
          '404':
            description: Evento não encontrado
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErroResponse'
          '500':
            description: Erro interno do servidor
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErroResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          example: "usuario@exemplo.com"
        password:
          type: string
          minLength: 8
          example: "Senha@123"
        name:
          type: string
          minLength: 3
          example: "João Silva"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "usuario@exemplo.com"
        password:
          type: string
          example: "Senha@123"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token para autenticação
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1c3VhcmlvQGV4ZW1wbG8uY29tIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        picture:
          type: string
          nullable: true
        accessLevel:
          type: string
          enum: [ADMIN, ADOTANTE, ONG]
...
    EventoResponseDto:
      type: object
      description: Dados de resposta de um evento
      required:
        - id
        - titulo
        - descricao
        - dataEvento
        - local
        - ongId
        - ongNome
        - statusEvento
        - dataCriacao
        - dataAtualizacao
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do evento
          example: "550e8400-e29b-41d4-a716-446655440000"
        titulo:
          type: string
          minLength: 3
          maxLength: 200
          description: Título do evento
          example: "Adoção em Indaiatuba"
        descricao:
          type: string
          minLength: 10
          maxLength: 2000
          description: Descrição detalhada do evento
          example: "Evento de adoção de animais no parque ecológico com atividades para as crianças"
        dataEvento:
          type: string
          format: date-time
          description: Data e hora do evento (ISO 8601)
          example: "2025-11-20T10:00:00Z"
        local:
          type: string
          minLength: 5
          maxLength: 300
          description: Local onde o evento será realizado
          example: "Parque Ecológico, Indaiatuba"
        ongId:
          type: string
          format: uuid
          description: ID da ONG criadora do evento
          example: "8a8a8a8a-8a8a-8a8a-8a8a-8a8a8a8a8a8a"
        ongNome:
          type: string
          description: Nome da ONG criadora do evento
          example: "Abrigo Amigos"
        statusEvento:
          type: string
          enum:
            - ATIVO
            - CANCELADO
            - FINALIZADO
            - ADIADO
          description: Status atual do evento
          example: "ATIVO"
        dataCriacao:
          type: string
          format: date-time
          description: Data e hora de criação do evento (ISO 8601)
          example: "2025-10-15T14:30:00Z"
        dataAtualizacao:
          type: string
          format: date-time
          description: Data e hora da última atualização do evento (ISO 8601)
          example: "2025-10-15T14:30:00Z"
    CriarEventoDto:
      type: object
      description: Dados para criar um novo evento
      required:
        - titulo
        - descricao
        - dataEvento
        - local
      properties:
        titulo:
          type: string
          minLength: 3
          maxLength: 200
          description: Título do evento
          example: "Adoção em Indaiatuba"
        descricao:
          type: string
          minLength: 10
          maxLength: 2000
          description: Descrição detalhada do evento
          example: "Grande evento de adoção com mais de 50 animais disponíveis"
        dataEvento:
          type: string
          format: date-time
          description: Data e hora do evento (ISO 8601)
          example: "2025-12-10T10:00:00Z"
        local:
          type: string
          minLength: 5
          maxLength: 300
          description: Local onde o evento será realizado
          example: "Praça, Indaiatuba"
    ErroResponse:
      type: object
      description: Resposta de erro padrão da API
      required:
        - mensagem
        - status
        - timestamp
      properties:
        mensagem:
          type: string
          description: Mensagem descritiva do erro
          example: "Campo 'titulo' é obrigatório"
        status:
          type: integer
          description: Código HTTP de status do erro
          example: 400
        timestamp:
          type: string
          format: date-time
          description: Data e hora do erro (ISO 8601)
          example: "2025-10-20T10:35:00Z"
        detalhes:
          type: string
          description: Detalhes adicionais do erro (opcional)
          example: "Erro ao processar a requisição"
        campoErro:
          type: string
          description: Campo específico com erro (opcional)
          example: "titulo"